Add tests for the externals service
